drivers/nble: Update Nordic RPC to 0518 release
Fixes limited discov mode bug and GATT write long request.
1 mandatory and 9 optional PTS test cases.
Fixes: RTOS-1524
Fixes: RTOS-1551

+++ drivers/bluetooth/nble/rpc_functions_to_ble_core.h
@@ -78,7 +78,7 @@
 		   uint8_t *, uint16_t)					\
 	FN_SIG_S_B(nble_gattc_write_req,				\
 		   const struct nble_gattc_write_req *,			\
-		   const uint8_t *, uint8_t)				\
+		   const uint8_t *, uint16_t)				\
 	FN_SIG_S_B(nble_gattc_read_multi_req,				\
 		   const struct nble_gattc_read_multi_req *,		\
 		   const uint16_t *, uint16_t)
